feat(hero): refactor Hero component and remove unused Hero2 component
